extract SeedAndPath in dedicated component
